Draft: Pre-Industrial, Custodian "Specializations"
Better settle the current situation surrounding the Custodians having a large issue about ranged combat. Since the Faction is, indeed, a non-ranged weapon wielding faction (it does not fit their niche), adjustments by choice of the hearthcore has became something I've promised. But since some people left, I had to do it myself with the help of beloved people.

However, I still see it a project made exactly as a "precursor" of the "Industrial Materiautism" project, and thus, will be nerfed or buffed as necessity calls for it.